Fix sliding sync on workers
